Remove the selection proxy object
I've replaced the selection proxy object, and replaced it with a simple
clone. The proxy object wasn't really pulling its weight, and it was
locking us into having a selector at the beginning of every method
signature, which made things like #first impossible

This PR also removes the optional selector (e.g. `.text('h1', 'foo')`)
as an argument to all of the various methods, in favor of
`.find('h1').text('foo')`